+++ crates/eww/src/config/inbuilt.rs
@@ -6,7 +6,7 @@ use yuck::config::{
     var_definition::VarDefinition,
 };
 
-use crate::{config::system_stats::*, EwwPaths};
+use crate::{config::system_stats::*, paths::EwwPaths};
 use eww_shared_util::VarName;
 
 macro_rules! define_builtin_vars {
@@ -18,7 +18,6 @@ macro_rules! define_builtin_vars {
                 VarName::from($name) => ScriptVarDefinition::Poll(PollScriptVar {
                     name: VarName::from($name),
                     run_while_expr: SimplExpr::Literal(DynVal::from(true)),
-                    run_while_var_refs: Vec::new(),
                     command: VarSource::Function($fun),
                     initial_value: None,
                     interval: $interval,