+++ crates/yuck/src/lib.rs
@@ -2,6 +2,7 @@
 #![allow(unused)]
 #![feature(try_blocks)]
 
+pub mod ast_error;
 pub mod config;
 pub mod error;
 pub mod format_diagnostic;